requirements.txt
Pillow==7.1.2
pyperclip==1.8.0
PyQt5==5.14.2
PyQt5-sip==12.7.2
pytesseract==0.3.4
py-notifier==0.1.0
win10toast==0.9; platform_system == "Windows"
googletrans==3.0.0